menambahkan terjual dan subtotal

+++ hapus.php
@@ -1,13 +1,18 @@
 <?php
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
 require_once 'koneksi.php';
 
-$id = $_GET['id'];
-$sql = "DELETE FROM barang WHERE id_barang=$id";
-
-if (mysqli_query($conn, $sql)) {
-    header("Location: index.php");
-    exit;
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
+    $sql = "DELETE FROM barang WHERE id_barang=$id";
+    if (mysqli_query($conn, $sql)) {
+        header("Location: index.php");
+        exit;
+    } else {
+        echo "Query Error: " . mysqli_error($conn);
+    }
 } else {
-    echo "Error: " . mysqli_error($conn);
+    echo "ID tidak ditemukan.";
 }
 ?>
